Allow summary to be optional
When viewing the feeds in some readers the summary (non rendered and
truncated text kinda sucks. Seems they prefer the summary and ignore the
content. But if summary is not there the content is rendered beautifully
and the readers manage their own truncation/show/hiding. Win Win in my
book. Also it still can show a text summary if the tiddler has a summary
field.