Convert gulp to make
Prior to this change, node was held back to <10 to accommodate gulp 3
since gulp 4 is not released in NPM yet. Attempts to upgrade to
gulp@next meant significant rewrites of the gulptfile. Since I'd have to
rewrite most of it it was just as easy to port this to a Makefile.