Add comment icon to toolbars (with scrolling)
Prior to this change, the user had to scroll to the bottom of the
article to view comments.

This change adds an icon to the toolbar so that the user can also click
the comment icon and it will scroll to the comment section and open the
Disqus comments. It uses the disqus-did-insert-element event from
tw5-discus PR #2