Add matcher support to array-updater
Prior to this change, the expectation was the comparison of array data
was equality. This forced the array being compared to be the original
(mutated) but there are cases when we may want an immutable array.

This change add the support for a custom matcher which we can use to
compare other aspects of the data (maybe item.id) instead of object
(reference) equality.