Upgrade validity.js
Prior to this change, we could only assign one validator per element. We
could attach multiple event handlers to the same event. And we lost the
full set of errors (if ever we needed it).

This change fixes all of this. And I think cleans up the code and
tests.