Disable rewrite for /fiddles
Prior to this change, 404 in the fiddles would rewrite and it would load
the main tiddlywiki into the fiddle instead of letting the fiddle code
handle the 404 error.

This change stops rewrite happening in the fiddle subtree. We needed
a missing.html to be available so that the 404 response worked in
fiddles ― otherwise the missing missing.html would trigger the top level
303 redirect rewrite.